+++ cmd/hidden_lake/adapters/chatingar/pkg/settings/settings.go
@@ -1,0 +1,16 @@
+package settings
+
+const (
+	CServiceName     = "HLA=chatingar"
+	CServiceFullName = "hidden-lake-adapters=chatingar"
+)
+
+const (
+	CPathYML = "hla_chatingar.yml"
+)
+
+const (
+	CDefaultHTTPAddress = "127.0.0.1:9531"
+	CDefaultPostID      = "65f7214f5b65dcbdedcca3fb"
+	CDefaultWaitTimeMS  = 5_000 // 5 seconds
+)