+++ cmd/hidden_lake/adapters/common/pkg/settings/settings.go
@@ -1,0 +1,17 @@
+package settings
+
+const (
+	CServiceName     = "HLA=common"
+	CServiceFullName = "hidden-lake-adapters=common"
+)
+
+const (
+	CPathYML = "hla_common.yml"
+	CPathDB  = "hla_common.db"
+)
+
+const (
+	CDefaultHTTPAddress = "127.0.0.1:9531"
+	CDefaultSrvAddress  = "127.0.0.1:6060"
+	CDefaultWaitTimeMS  = 1_000 // 1 second
+)