+++ cmd/hidden_lake/composite/pkg/app/init_app.go
@@ -4,31 +4,33 @@ import (
 	"path/filepath"
 	"strings"
 
+	"github.com/number571/go-peer/cmd/hidden_lake/composite/internal/config"
+	"github.com/number571/go-peer/cmd/hidden_lake/composite/pkg/settings"
 	"github.com/number571/go-peer/internal/flag"
 	"github.com/number571/go-peer/pkg/types"
 	"github.com/number571/go-peer/pkg/utils"
 
-	hla_chatingar_consumer_app "github.com/number571/go-peer/cmd/hidden_lake/adapters/chatingar/consumer/pkg/app"
-	hla_chatingar_consumer_settings "github.com/number571/go-peer/cmd/hidden_lake/adapters/chatingar/consumer/pkg/settings"
-	hla_chatingar_producer_app "github.com/number571/go-peer/cmd/hidden_lake/adapters/chatingar/producer/pkg/app"
-	hla_chatingar_producer_settings "github.com/number571/go-peer/cmd/hidden_lake/adapters/chatingar/producer/pkg/settings"
-	hla_common_consumer_app "github.com/number571/go-peer/cmd/hidden_lake/adapters/common/consumer/pkg/app"
-	hla_common_consumer_settings "github.com/number571/go-peer/cmd/hidden_lake/adapters/common/consumer/pkg/settings"
-	hla_common_producer_app "github.com/number571/go-peer/cmd/hidden_lake/adapters/common/producer/pkg/app"
-	hla_common_producer_settings "github.com/number571/go-peer/cmd/hidden_lake/adapters/common/producer/pkg/settings"
+	hla_chatingar_app "github.com/number571/go-peer/cmd/hidden_lake/adapters/chatingar/pkg/app"
+	hla_chatingar_settings "github.com/number571/go-peer/cmd/hidden_lake/adapters/chatingar/pkg/settings"
+
+	hla_common_app "github.com/number571/go-peer/cmd/hidden_lake/adapters/common/pkg/app"
+	hla_common_settings "github.com/number571/go-peer/cmd/hidden_lake/adapters/common/pkg/settings"
 
 	hlf_app "github.com/number571/go-peer/cmd/hidden_lake/applications/filesharer/pkg/app"
 	hlf_settings "github.com/number571/go-peer/cmd/hidden_lake/applications/filesharer/pkg/settings"
+
 	hlm_app "github.com/number571/go-peer/cmd/hidden_lake/applications/messenger/pkg/app"
 	hlm_settings "github.com/number571/go-peer/cmd/hidden_lake/applications/messenger/pkg/settings"
-	"github.com/number571/go-peer/cmd/hidden_lake/composite/internal/config"
-	"github.com/number571/go-peer/cmd/hidden_lake/composite/pkg/settings"
+
 	hle_app "github.com/number571/go-peer/cmd/hidden_lake/helpers/encryptor/pkg/app"
 	hle_settings "github.com/number571/go-peer/cmd/hidden_lake/helpers/encryptor/pkg/settings"
+
 	hll_app "github.com/number571/go-peer/cmd/hidden_lake/helpers/loader/pkg/app"
 	hll_settings "github.com/number571/go-peer/cmd/hidden_lake/helpers/loader/pkg/settings"
+
 	hlt_app "github.com/number571/go-peer/cmd/hidden_lake/helpers/traffic/pkg/app"
 	hlt_settings "github.com/number571/go-peer/cmd/hidden_lake/helpers/traffic/pkg/settings"
+
 	hls_app "github.com/number571/go-peer/cmd/hidden_lake/service/pkg/app"
 	hls_settings "github.com/number571/go-peer/cmd/hidden_lake/service/pkg/settings"
 )
@@ -71,14 +73,10 @@ func getRunners(pCfg config.IConfig, pArgs []string, pDefaultPath, pDefaultKey s
 			runner, err = hlm_app.InitApp(pArgs, pDefaultPath)
 		case hlf_settings.CServiceFullName:
 			runner, err = hlf_app.InitApp(pArgs, pDefaultPath)
-		case hla_common_consumer_settings.CServiceFullName:
-			runner, err = hla_common_consumer_app.InitApp(pArgs, pDefaultPath)
-		case hla_common_producer_settings.CServiceFullName:
-			runner, err = hla_common_producer_app.InitApp(pArgs, pDefaultPath)
-		case hla_chatingar_consumer_settings.CServiceFullName:
-			runner, err = hla_chatingar_consumer_app.InitApp(pArgs, pDefaultPath)
-		case hla_chatingar_producer_settings.CServiceFullName:
-			runner, err = hla_chatingar_producer_app.InitApp(pArgs, pDefaultPath)
+		case hla_common_settings.CServiceFullName:
+			runner, err = hla_common_app.InitApp(pArgs, pDefaultPath)
+		case hla_chatingar_settings.CServiceFullName:
+			runner, err = hla_chatingar_app.InitApp(pArgs, pDefaultPath)
 		default:
 			return nil, ErrUnknownService
 		}